resolve coding convention

+++ app/Http/Controllers/Admin/QuestionController.php
@@ -42,39 +42,40 @@ class QuestionController extends Controller
         // @phpstan-ignore-next-line
         return DataTables::of($questions)
         ->editColumn('course_id', function ($question) {
-            $courseName = $question->course->title;
-            return $courseName;
+            return $question->course->title;
         })
         ->editColumn('category', function ($question) {
-            if($question->category == 0) return 'Câu hỏi tự luận';
-            if($question->category == 1) return 'Câu hỏi nhiều lựa chọn';
-            if($question->category == 2) return 'Câu hỏi đúng sai';
-            if($question->category == 3) return 'Câu hỏi trắc nghiệm';
+            switch ($question->category){
+                case 0:
+                    return 'Câu hỏi tự luận';
+                case 1:
+                    return 'Câu hỏi nhiều lựa chọn';
+                case 2:
+                    return 'Câu hỏi đúng sai';
+                case 3:
+                    return 'Câu hỏi trắc nghiệm';
+                default:
+                    return '';
+            }
         })
         ->addColumn('answers', function ($question) {
-            if ($question->category == 1){
-                $var = <<<EOD
-                <a onclick="event.preventDefault();answer_qu($question->id)" 
+            $tmp = <<<EOD
+                <a onclick="event.preventDefault();answer_qu($question->id)"
                     href="" class="btn btn-primary btn-sm " title="Xem câu trả lời">
                     <i class="fa fa-plus-circle"></i></a>
-                EOD;
-                return $var;
+EOD;
+            switch($question->category){
+                case 0:
+                    return 'Tự luận';
+                case 1:
+                    return $tmp;
+                case 2:
+                    return (1 == $question->answer)?'Đúng':'Sai';
+                case 3:
+                    return $tmp;
+                default:
+                    return '';
             }
-            else 
-            if ($question->category == 2){
-                if($question->answer == 1) return 'Đúng';
-                return 'Sai';
-            }
-            elseif ($question->category == 3){
-                $var = <<<EOD
-                <a onclick="event.preventDefault();answer_qu($question->id)" 
-                    href="" class="btn btn-primary btn-sm " title="Xem câu trả lời">
-                    <i class="fa fa-plus-circle"></i></a>
-                EOD;
-                return $var;
-            }
-            else return 'Tự luận';
-
         })
         ->addColumn('actions', function ($question) {
             return view('admin.questions.actions', ['row' => $question])->render();
@@ -154,46 +155,48 @@ class QuestionController extends Controller
                     'answer' => $question_item['answer2']
                 ]);
             }
-        } catch (Exception $t) {
+        } catch (\Exception $t) {
             DB::rollBack();
             // log error
-            throw new Exception($t->getMessage());
+            throw new \Exception($t->getMessage());
         }
         DB::commit();
-        return redirect(route('question.index'))->with('message', 'Thêm câu hỏi thành công !')->with('type_alert', "success");
+        return redirect(route('question.index'))
+        ->with('message', 'Thêm câu hỏi thành công !')
+        ->with('type_alert', "success");
     }
 
     /**
      * @param Request $request
      * @param int $id
-     * @return \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
+     *
+     * @return \Illuminate\Contracts\View\View|\Illuminate\View\View|\Illuminate\Http\RedirectResponse
      */
     public function edit(Request $request, $id)
     {
-        $question_test = Question::find($id);
-        if ($question_test->tests()->exists()) {
-            return redirect(route('question.index'))
+        $question = Question::find($id);
+        if ($question){
+            if ($question->tests()->exists()) {
+                return redirect(route('question.index'))
                 ->with('message', "Câu hỏi có trong bài test không thể sửa !")
                 ->with('type_alert', "danger");
-        } else {
-            $question = Question::find($id);
-            if ($question) {
-
+            }else{
                 $answers = Answer::where('question_id', $id)->get();
                 $course = Course::all();
 
-                return view('admin.questions.edit', compact([
+                return view('admin.questions.edit', compact(
                     'question',
                     'answers',
                     'course'
-                ]));
+                    ));
             }
         }
+        return abort(404);
     }
 
     /**
      * @param QuestionRequest $request
-     *
+     * @param integer $id
      * @return \Illuminate\Routing\Redirector|\Illuminate\Http\RedirectResponse
      */
     public function update(QuestionRequest $request, $id)
@@ -231,7 +234,9 @@ class QuestionController extends Controller
             $msg = 'Sửa thành công câu hỏi :' . $question->content;
         }
 
-        return redirect(route('question.index'))->with('message', $msg)->with('type_alert', "success");
+        return redirect(route('question.index'))
+        ->with('message', $msg)
+        ->with('type_alert', "success");
     }
 
     /**
@@ -241,23 +246,23 @@ class QuestionController extends Controller
      */
     public function destroy(Request $request)
     {
-        $question_id = $request->input('question_id', 0);
+        $question_id = 0 + $request->input('question_id', 0);
         $question = Question::find($question_id);
-
-        if ($question->tests()->exists()) {
-            return redirect(route('question.index'))
-                ->with('message', "Câu hỏi có trong bài test không thể xóa !")
-                ->with('type_alert', "danger");
-        } else {
-            if ($question_id) {
-                Question::destroy($question_id);
+        if($question){
+            if ($question->tests()->exists()) {
                 return redirect(route('question.index'))
-                    ->with('message', "Xóa câu hỏi {$question_id} thành công !")
-                    ->with('type_alert', "success");
+                    ->with('message', "Câu hỏi có trong bài test không thể xóa !")
+                    ->with('type_alert', "danger");
             } else {
-                throw new ModelNotFoundException();
+                if ($question_id) {
+                    Question::destroy($question_id);
+                    return redirect(route('question.index'))
+                        ->with('message', "Xóa câu hỏi {$question_id} thành công !")
+                        ->with('type_alert', "success");
+                }
             }
         }
+        throw new ModelNotFoundException();
     }
 
     /**
@@ -268,9 +273,8 @@ class QuestionController extends Controller
     {
         $output = '';
         $answers = Answer::where('question_id', $id)->get();
-        // dd($id);
-        if ($answers) {
 
+        if ($answers->count()) {
             foreach ($answers as $an) {
                 if ($an->checked == 1) {
                     $checked = 'Đúng';
